Day 14 of AoC 2021
Two different methods for reaching the same objective. The first iterates over the array and inserts new elements. This iterative approach is big O(n) and doesn't scale with increasing steps. The second solution is linear, and counts items as they are added. This solution is very fast and scales well.